requirements.txt
Flask
zaproxy
gunicorn
python-dotenv